requirements.txt
flet==0.24.1
pillow
requests
